Added back fade-out in credits. Fixed particles hanging on left side, not leaving screen Refactorization: TColor3 uses uint8_t instead of double.

+++ branches/SFML2/src/etr_types.h
@@ -26,8 +26,8 @@ enum Orientation {
 };
 
 struct TColor3 {
-	double r, g, b;
-	TColor3(double r_ = 0, double g_ = 0, double b_ = 0)
+	uint8_t r, g, b;
+	TColor3(uint8_t r_ = 0, uint8_t g_ = 0, uint8_t b_ = 0)
 		: r(r_), g(g_), b(b_)
 	{}
 };
